+++ packages/transactional/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @folionote/transactional
 
+## 0.2.10
+
+### Patch Changes
+
+- feat: ✨ chat message persistence
+
 ## 0.2.9
 
 ### Patch Changes